src/backend/requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
python-multipart
asyncpg
redis
python-dotenv
python-jose[cryptography]
uuid
databases
pydantic-settings
openai
tenacity
pytest
pytest-asyncio
httpx
mock
aiohttp>=3.8.6
email_validator
bcrypt
pyjwt
requests>=2.31.0
tiktoken>=0.5.1
sentence-transformers
pymilvus
PyPDF2>=3.0.0
langchain>=0.1.0
unstructured[pdf,doc,excel,image,html]==0.12.4
elasticsearch
scikit-learn
ragas 